requirements.txt
asgiref==3.7.2
Django==4.2.7
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2023.3
asgiref==3.7.2
Django==4.2.7
django-appconf==1.0.5
et-xmlfile==1.1.0
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
Pillow==10.1.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
